requirements-azure.txt
azure-cli==2.34.1
azure-storage-blob==12.9.0